requirements.txt
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
